+++ frontend/src/Page/CreateMessagePage/index.js
@@ -1,135 +1,9 @@
 import React, { useState } from "react";
 import { CgAdd } from "react-icons/cg";
-import { BsTrash } from "react-icons/bs";
-import { AiFillHome, AiOutlineClose, AiOutlineMessage, AiOutlineLink, AiOutlinePicture } from "react-icons/ai";
-import "../style.css";
+import "../CreateMessagePage/style.css";
+import MainModal from "../../Components/MainModal/MainModal";
+import MessageRevisedModal from "../../Components/MessageReviseModal/MessageReviseModal";
 
-function MessageModal({ closeMessageModalFunc, registerMessageFunc, closeModalFunc }) {
-    const [message, setMessage] = useState("");
-    const cancelClicked = () => {
-        closeMessageModalFunc();
-    };
-    const textareaHandler = (e) => {
-        setMessage(e.target.value);
-    };
-    const registerMessage = () => {
-        registerMessageFunc(message);
-        closeMessageModalFunc();
-        closeModalFunc();
-    };
-
-    return (
-        <div className="message-modal">
-            <div className="header">
-                <div className="title">하고 싶은 말을 적어보세요</div>
-            </div>
-            <div className="body">
-                <div className="color-container"></div>
-                <div className="message-container">
-                    <textarea
-                        className="message"
-                        placeholder="내용을 입력하세요!"
-                        onChange={textareaHandler}
-                    ></textarea>
-                </div>
-            </div>
-            <div className="footer">
-                <div className="cancel" onClick={cancelClicked}>
-                    취소
-                </div>
-                <div className="register" onClick={registerMessage}>
-                    등록
-                </div>
-            </div>
-        </div>
-    );
-}
-function MessageRevisedModal({ cloud, reviseCloudFunc, closeModalFunc, removeCloudFunc }) {
-    const [message, setMessage] = useState(cloud.message);
-    const textareaHandler = (e) => {
-        setMessage(e.target.value);
-    };
-    const reviseCloud = () => {
-        reviseCloudFunc(cloud, message);
-        closeModalFunc();
-    };
-    const closeModal = () => {
-        closeModalFunc();
-    };
-    const removeCloud = () => {
-        removeCloudFunc(cloud.id);
-        closeModalFunc();
-    };
-
-    return (
-        <div className="message-modal">
-            <div className="header" style={{ backgroundColor: "red", borderColor: "red", opacity: "0.8" }}>
-                <BsTrash color="black" size="32" onClick={removeCloud}></BsTrash>
-            </div>
-            <div className="body">
-                <div className="color-container"></div>
-                <div className="message-container">
-                    <textarea className="message" defaultValue={message} onChange={textareaHandler}></textarea>
-                </div>
-            </div>
-            <div className="footer">
-                <div className="cancel" onClick={closeModal}>
-                    취소
-                </div>
-
-                <div className="register" onClick={reviseCloud}>
-                    수정
-                </div>
-            </div>
-        </div>
-    );
-}
-
-function MainModal({ closeModalFunc, addCloudFunc }) {
-    const [messageModal, setMessageModal] = useState(false);
-    const [mainModal, setMainModal] = useState(true);
-    const closeModal = () => {
-        closeModalFunc();
-    };
-    const showMessageModal = () => {
-        setMessageModal(true);
-        setMainModal(false);
-    };
-    const registerMessage = (message) => {
-        addCloudFunc(message);
-    };
-    const closeMessageModal = () => {
-        setMessageModal(false);
-        setMainModal(true);
-    };
-
-    return (
-        <>
-            {" "}
-            {messageModal === true ? (
-                <MessageModal
-                    closeMessageModalFunc={closeMessageModal}
-                    registerMessageFunc={registerMessage}
-                    closeModalFunc={closeModalFunc}
-                    isNew={true}
-                />
-            ) : null}
-            {mainModal === true ? (
-                <div className="modal">
-                    <div className="header">
-                        <div className="title">추가 하기</div>
-                        <AiOutlineClose color="#ff97bf" onClick={closeModal}></AiOutlineClose>
-                    </div>
-                    <div className="body">
-                        <AiOutlineMessage color="#ff97bf" onClick={showMessageModal} size="64"></AiOutlineMessage>
-                        <AiOutlinePicture color="#ff97bf" size="64"></AiOutlinePicture>
-                        <AiOutlineLink color="#ff97bf" size="64"></AiOutlineLink>
-                    </div>
-                </div>
-            ) : null}
-        </>
-    );
-}
 function CreateMessagePage() {
     const [modal, setModal] = useState(false);
