+++ frontend/src/Page/MainPage/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../style.css";
+import "./style.css";
 import { useNavigate } from "react-router-dom";
 function MainPage() {
     const next = () => {
